backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pyjwt
passlib[bcrypt]
bcrypt
